backend/requirements.txt
# Core FastAPI Framework
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6

# Document Processing (Resume Parsing)
PyPDF2>=3.0.0
python-docx>=1.1.0
pdfplumber>=0.10.0  # Better PDF extraction than PyPDF2

# Data Processing & ML
pydantic>=2.0.0
pydantic-settings>=2.0.0
scikit-learn>=1.3.0
numpy>=1.24.0
pandas>=2.0.0

# Configuration & Environment
python-dotenv>=1.0.0

# HTTP & API Clients
requests>=2.31.0
httpx>=0.25.0  # Async HTTP client for better performance

# Email & Communication
email-validator>=2.1.0
dnspython>=2.4.0
imapclient>=2.3.1  # Better IMAP handling

# Microsoft Authentication
msal>=1.24.0

# Database
aiosqlite>=0.19.0  # Async SQLite for better concurrency

# AI Models - LOCAL (FREE, ZERO API COSTS)
sentence-transformers>=2.2.0  # Semantic similarity AI
torch>=2.0.0  # Required by sentence-transformers
transformers>=4.35.0  # Additional NLP capabilities
gpt4all>=2.0.0  # Optional: Full local LLM (~4GB)

# Intel GPU Acceleration (Intel Iris Xe)
# intel-extension-for-pytorch>=2.1.0  # Intel GPU optimization - install separately with: pip install intel-extension-for-pytorch
# openvino>=2023.2.0  # Intel OpenVINO for fast inference - optional

# Named Entity Recognition (NER) for accurate extraction
spacy>=3.7.0  # NLP pipeline for NER
# Download model: python -m spacy download en_core_web_sm

# AI - EMERGENCY FALLBACK ONLY (costs money)
openai>=1.0.0  # Only used on timeout/error

# Performance & Optimization
cachetools>=5.3.0  # Advanced caching
redis>=5.0.0  # Optional: Redis cache for production
python-lru-cache>=0.1.0  # LRU caching

# Async & Concurrency
asyncio>=3.4.3
aiofiles>=23.2.0  # Async file operations
aiohttp>=3.9.0  # Async HTTP client for calendar/SMS APIs

# SMS & Communication
twilio>=8.0.0  # Twilio SMS integration

# Monitoring & Logging
python-json-logger>=2.0.0  # Structured logging
psutil>=5.9.0  # System resource monitoring

# Security
python-jose[cryptography]>=3.3.0  # JWT tokens
passlib[bcrypt]>=1.7.4  # Password hashing
cryptography>=41.0.0

# Compression
brotli>=1.1.0  # Brotli compression (optional, gzip built-in)

# Development & Testing
watchfiles>=0.21.0  # Auto-reload (included in uvicorn[standard])
pytest>=7.4.0
pytest-asyncio>=0.21.0
httpx>=0.25.0  # For testing async endpoints

